Add gas measurement for `bulkClaim`
What's the right amount of IOUs to use for that? Or should I do
measurements with multiple different lengths? I just went with a single
IOU for now.
