+++ raiden_contracts/data/gas.json
@@ -4,20 +4,21 @@
     "CustomToken.transfer": 36123,
     "CustomToken.transferFrom": 29557,
     "MonitoringService.claimReward": 41942,
-    "MonitoringService.monitor": 207777,
+    "MonitoringService.monitor": 207760,
+    "OneToN.bulkClaim": 81653,
     "OneToN.claim": 92786,
     "SecretRegistry.registerSecret": 46892,
     "ServiceRegistry.deposit": 527632,
     "ServiceRegistry.setURL": 44676,
-    "TokenNetwork DEPLOYMENT": 4357148,
-    "TokenNetwork.closeChannel": 124163,
-    "TokenNetwork.openChannel": 97755,
-    "TokenNetwork.setTotalDeposit": 44919,
-    "TokenNetwork.setTotalWithdraw": 103318,
-    "TokenNetwork.settleChannel": 108528,
-    "TokenNetwork.unlock 1 locks": 31539,
-    "TokenNetwork.unlock 6 locks": 58595,
-    "TokenNetwork.updateNonClosingBalanceProof": 94043,
+    "TokenNetwork DEPLOYMENT": 4357212,
+    "TokenNetwork.closeChannel": 124035,
+    "TokenNetwork.openChannel": 97745,
+    "TokenNetwork.setTotalDeposit": 44855,
+    "TokenNetwork.setTotalWithdraw": 103237,
+    "TokenNetwork.settleChannel": 108432,
+    "TokenNetwork.unlock 1 locks": 31475,
+    "TokenNetwork.unlock 6 locks": 58531,
+    "TokenNetwork.updateNonClosingBalanceProof": 93979,
     "TokenNetworkRegistry DEPLOYMENT": 5027825,
     "TokenNetworkRegistry createERC20TokenNetwork": 3360447,
     "UserDeposit.deposit": 92392,
